SHA-3 Python Program, My Final Update For Now
I made this following documentation that I posted in the resources tab. Please note that it lacks the padding step that needs to be implemented but it should work fine as the core algorithm for the SHA-3 component of our project with a little work. Please test it out and let me know if it is working right, Python is acting a fool on my end once more. If any logic needs to be changed it is likely the axis logic for the arrays but this should work fine for generating hashes.